pyproject.toml
[tool.poetry]
name = "ska-csp-simulators"
version = "0.1.0"
description = "SKA CSP Simulators"
authors = ["Elisabetta Giani <elisabetta.giani@inaf.it>"]
license = "BSD-3-Clause"
packages = [
    { include = "ska_csp_simulators", from = "src" }
]

[tool.poetry.scripts]
BaseSimulatorDevice = 'ska_csp_simulators.common.BaseSimulatorDevice:main'

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
debugpy = "^1.5.1"
numpy = "^1.26.1"
pytango = "^9.5.0"
ska-control-model= '0.3.1'
ska-k8s-config-exporter = "^0.0.4"
ska-ser-log-transactions = "*"
ska-tango-testing = "0.6.0"
notebook = "^7.0.6"
jupyterlab = "^4.0.8"
jupyterhub = "^4.0.2"
jupyter-server = "^2.10.0"

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
coverage = "^6.1.1"
flake8 = "^3.9.2"
isort = "^5.6.4"
nbmake = {extras = ["all"], version = "^1.4.1"}
nbqa = {extras = ["all"], version = "^1.7.0"}
pylint = "^2.8.2"
pylint-junit = "^0.3.2"
pytest = "^7.1.0"
pytest-bdd = "^4.0.1"
pytest-cov = "^2.10.1"
pytest-forked = "^1.3.0"
pytest-json-report = "^1.4.1"
pytest-mock = "^3.6.1"
pytest-repeat = "^0.9.1"
pytest-timeout = "^2.0.1"
yamllint = "^1.32.0"

[tool.poetry.group.docs.dependencies]
myst-parser = "^2.0.0"
ska-ser-sphinx-theme = "^0.1.1"
Sphinx = "^6"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
